I've added request as a dependency, since this was the easiest way to get it working behind a corporate proxy. I've also moved the code around, separated the command line stuff from the main code and added a way of turning the debug messages on and off. I separated the code to make it a little easier to write tests. I also added tests, and set up package.json to allow "npm test" to run them. Dev dependencies now include nock (for intercepting http calls), mocha (for the test framework), and supertest (which makes testing http servers nice and easy).
The delegate itself now makes two requests to a remote repository. First is a HEAD, to see if the resource exists,
followed by a GET if it does. I do this so that I can pipe the GET response to the caller easily - I inspect the
status code from the HEAD, but to do that in the GET would mean that request would read and buffer the whole response
before calling my callback. This way I get to use the nifty request pipe support, which handles all the headers and things.